Use 32-bit byte indices, 32-bit prime counter when sieving
The index of the byte to mark should never required 64 bits (as that
would mean a very long sieve bit array!), so locally use only 32 bits to
keep track of this value.  This produces a big performance boost and
32-bit systems.

Additionally, use a local counter in sieve_segment() and only update the
passed counter once.  On 32-bit systems this removes the overhead of
having to add repeatedly to a uint64_t.

+++ include/yase.h
@@ -215,7 +215,7 @@ void prime_set_cleanup(struct prime_set * set);
 static inline void mark_multiple_30(
 		uint8_t *  sieve,
 		uint32_t   prime_adj,
-		uint64_t * byte,
+		uint32_t * byte,
 		uint32_t * wheel_idx)
 {
 	sieve[*byte] |= wheel30[*wheel_idx].mask;
@@ -229,7 +229,7 @@ static inline void mark_multiple_30(
 static inline void mark_multiple_210(
 		uint8_t *  sieve,
 		uint32_t   prime_adj,
-		uint64_t * byte,
+		uint32_t * byte,
 		uint32_t * wheel_idx)
 {
 	sieve[*byte] |= wheel210[*wheel_idx].mask;
